Create simple_network_sim package
